Update the jwt integration.
