+++ src/UserBundle/Resources/config/security.php
@@ -24,6 +24,7 @@ return function (ContainerConfigurator $container): void {
         ->alias(PasswordHashingInterface::class, Security\PasswordHashing::class)
 
         ->set(Security\SecurityUserProvider::class)
+        ->set(Security\Jwt\SecurityUserProvider::class)
         ->set(Security\UserValueResolver::class)
             ->tag('controller.argument_value_resolver')
     ;
